+++ GCBot/GCBot.Models/Backup/UserReport.cs
@@ -5,7 +5,7 @@ namespace GCBot.Models.Backup
 {
     public class UserReport : Report<DateTime, int>
     {
-        public UserReport(uint userId, DateRange dateRange) : base(dateRange,userId) { }
+        public UserReport(ulong userId, DateRange dateRange) : base(dateRange,userId) { }
 
         public override int TotalMessages =>
             Information.Values.Sum(a => a);
