+++ GCBot/GCBot.Services.EntityFramework/Entities/Message.cs
@@ -4,10 +4,10 @@ namespace GCBot.Services.EntityFramework.Entities
 {
     public class Message
     {
-        public uint Id { get; set; }
+        public int Id { get; set; }
         public DateTime DateSent { get; set; }
-        public uint SenderId { get; set; }
-        public uint ChannelId { get; set; }
+        public ulong SenderId { get; set; }
+        public ulong ChannelId { get; set; }
         public string Text { get; set; }
         public string Link { get; set; }
     }
